1、编译通过；2、增加非校准感应器支持；

+++ app/src/main/java/org.achartengine/chart/DoughnutChart.java
@@ -115,7 +115,7 @@ public class DoughnutChart extends RoundChart {
             mRenderer.getLabelsColor(), paint, true, false);
         currentAngle += angle;
       }
-      radius -= (int) mRadius * decCoef;
+      radius -= mRadius * decCoef;
       shortRadius -= mRadius * decCoef - 2;
       if (mRenderer.getBackgroundColor() != 0) {
         paint.setColor(mRenderer.getBackgroundColor());
